#165:usePopupの変更とただのこんぽーねんとになったので移動
